chore: disable routing changes to investigate inscription issues
