+++ src/app/pages/send/ordinal-inscription/components/send-inscription-container.tsx
@@ -39,6 +39,8 @@ export function SendInscriptionContainer() {
     setUtxo(createUtxoFromInscription(routeState.inscription));
   });
 
+  console.log({ inscription });
+
   if (!inscription || !utxo) return null;
 
   return (
