+++ src/app/store/accounts/blockchain/bitcoin/native-segwit-account.hooks.ts
@@ -1,36 +1,23 @@
 import { useMemo } from 'react';
 import { useSelector } from 'react-redux';
 
-import { PaymentTypes } from '@btckit/types';
 import { createSelector } from '@reduxjs/toolkit';
-import * as btc from '@scure/btc-signer';
 import { Psbt } from 'bitcoinjs-lib';
-import AppClient from 'ledger-bitcoin';
 
-import { getBitcoinJsLibNetworkConfigByMode } from '@shared/crypto/bitcoin/bitcoin.network';
 import {
   deriveAddressIndexZeroFromAccount,
-  getInputPaymentType,
   lookUpLedgerKeysByPath,
 } from '@shared/crypto/bitcoin/bitcoin.utils';
-import { getTaprootAccountDerivationPath } from '@shared/crypto/bitcoin/p2tr-address-gen';
 import {
   deriveNativeSegwitAccountFromRootKeychain,
   getNativeSegWitPaymentFromAddressIndex,
   getNativeSegwitAccountDerivationPath,
 } from '@shared/crypto/bitcoin/p2wpkh-address-gen';
-import { logger } from '@shared/logger';
 import { makeNumberRange, reverseBytes } from '@shared/utils';
 
 import { mnemonicToRootNode } from '@app/common/keychain/keychain';
-import {
-  addNativeSegwitSignaturesToPsbt,
-  addTaprootInputSignaturesToPsbt,
-  createNativeSegwitDefaultWalletPolicy,
-  createTaprootDefaultWalletPolicy,
-} from '@app/features/ledger/utils/bitcoin-ledger-utils';
 import { useBitcoinClient } from '@app/store/common/api-clients.hooks';
-import { selectCurrentNetwork, useCurrentNetwork } from '@app/store/networks/networks.selectors';
+import { selectCurrentNetwork } from '@app/store/networks/networks.selectors';
 import { selectCurrentAccountIndex } from '@app/store/software-keys/software-key.selectors';
 
 import { useCurrentAccountIndex } from '../../account';
@@ -42,7 +29,6 @@ import {
   bitcoinAddressIndexSignerFactory,
   useMakeBitcoinNetworkSignersForPaymentType,
 } from './bitcoin-signer';
-import { useUpdateLedgerSpecificTaprootInputPropsForAdddressIndexZero } from './taproot-account.hooks';
 
 const selectNativeSegwitAccountBuilder = bitcoinAccountBuilderFactory(
   deriveNativeSegwitAccountFromRootKeychain,
@@ -166,86 +152,5 @@ export function useUpdateLedgerSpecificNativeSegwitInputPropsForAdddressIndexZer
         ],
       });
     });
-  };
-}
-
-export function useSignLedgerTx() {
-  const accountIndex = useCurrentAccountIndex();
-  const network = useCurrentNetwork();
-
-  const updateNativeSegwitLedgerProps =
-    useUpdateLedgerSpecificNativeSegwitInputPropsForAdddressIndexZero();
-
-  const updateTaprootLedgerInputs = useUpdateLedgerSpecificTaprootInputPropsForAdddressIndexZero();
-
-  return async (app: AppClient, rawPsbt: Uint8Array) => {
-    const fingerprint = await app.getMasterFingerprint();
-
-    // BtcSigner not compatible with Ledger. Encoded format returns more terse
-    // version. BitcoinJsLib works.
-    const psbt = Psbt.fromBuffer(Buffer.from(rawPsbt), {
-      network: getBitcoinJsLibNetworkConfigByMode(network.chain.bitcoin.network),
-    });
-
-    const btcSignerPsbtClone = btc.Transaction.fromPSBT(psbt.toBuffer());
-
-    const inputByPaymentType = makeNumberRange(btcSignerPsbtClone.inputsLength).map(index => [
-      index,
-      getInputPaymentType(index, btcSignerPsbtClone.getInput(index), network.chain.bitcoin.network),
-    ]) as readonly [number, PaymentTypes][];
-
-    //
-    // Taproot
-    const taprootInputsToSign = inputByPaymentType
-      .filter(([_, paymentType]) => paymentType === 'p2tr')
-      .map(([index]) => index);
-
-    if (taprootInputsToSign.length) {
-      await updateTaprootLedgerInputs(psbt, fingerprint, taprootInputsToSign);
-
-      const taprootExtendedPublicKey = await app.getExtendedPubkey(
-        getTaprootAccountDerivationPath(network.chain.bitcoin.network, accountIndex)
-      );
-
-      const taprootPolicy = createTaprootDefaultWalletPolicy({
-        fingerprint,
-        accountIndex,
-        network: network.chain.bitcoin.network,
-        xpub: taprootExtendedPublicKey,
-      });
-
-      const taprootSignatures = await app.signPsbt(psbt.toBase64(), taprootPolicy, null);
-      addTaprootInputSignaturesToPsbt(psbt, taprootSignatures);
-    }
-
-    //
-    // Native Segwit
-    const nativeSegwitInputsToSign = inputByPaymentType
-      .filter(([_, paymentType]) => paymentType === 'p2wpkh')
-      .map(([index]) => index);
-
-    if (nativeSegwitInputsToSign.length) {
-      const nativeSegwitExtendedPublicKey = await app.getExtendedPubkey(
-        getNativeSegwitAccountDerivationPath(network.chain.bitcoin.network, accountIndex)
-      );
-
-      await updateNativeSegwitLedgerProps(psbt, fingerprint, nativeSegwitInputsToSign);
-
-      const nativeSegwitPolicy = createNativeSegwitDefaultWalletPolicy({
-        fingerprint,
-        accountIndex,
-        network: network.chain.bitcoin.network,
-        xpub: nativeSegwitExtendedPublicKey,
-      });
-
-      const nativeSegwitSignatures = await app.signPsbt(psbt.toBase64(), nativeSegwitPolicy, null);
-      logger.debug('Native Segwit signatures', nativeSegwitSignatures);
-
-      addNativeSegwitSignaturesToPsbt(psbt, nativeSegwitSignatures);
-    }
-
-    // Turn back into BtcSigner lib format
-    // REMINDER: this tx is not finalized
-    return btc.Transaction.fromPSBT(psbt.toBuffer());
   };
 }
